+++ tests/apps/search/test_forms_search_items.py
@@ -1,25 +1,27 @@
 """
 Tests for the course search form.
 """
+from unittest import mock
+
 from django.http.request import QueryDict
 from django.test import TestCase
-from django.test.utils import override_settings
 
+from richie.apps.core.defaults import ALL_LANGUAGES_DICT
 from richie.apps.search.forms import ItemSearchForm
 
 
-@override_settings(ALL_LANGUAGES_DICT={"fr": "French", "en": "English"})
+@mock.patch.dict(ALL_LANGUAGES_DICT, {"fr": "French", "en": "English"})
 class ItemSearchFormTestCase(TestCase):
     """
     Test the course search form.
     """
 
-    def test_forms_items_params_not_required(self):
+    def test_forms_items_params_not_required(self, *_):
         """No params are required for the search form."""
         form = ItemSearchForm()
         self.assertTrue(form.is_valid())
 
-    def test_forms_items_empty_querystring(self):
+    def test_forms_items_empty_querystring(self, *_):
         """The empty query string should be a valid search form."""
         form = ItemSearchForm(data=QueryDict())
         self.assertTrue(form.is_valid())
@@ -27,7 +29,7 @@ class ItemSearchFormTestCase(TestCase):
             form.cleaned_data, {"limit": None, "offset": None, "query": "", "scope": ""}
         )
 
-    def test_forms_courses_limit_greater_than_1(self):
+    def test_forms_courses_limit_greater_than_1(self, *_):
         """The `limit` param should be greater than 1."""
         form = ItemSearchForm(data=QueryDict(query_string="limit=0"))
         self.assertFalse(form.is_valid())
@@ -35,7 +37,7 @@ class ItemSearchFormTestCase(TestCase):
             form.errors, {"limit": ["Ensure this value is greater than or equal to 1."]}
         )
 
-    def test_forms_courses_limit_integer(self):
+    def test_forms_courses_limit_integer(self, *_):
         """The `limit` param should be an integer."""
         form = ItemSearchForm(data=QueryDict(query_string="limit=a"))
         self.assertFalse(form.is_valid())
@@ -44,7 +46,7 @@ class ItemSearchFormTestCase(TestCase):
         form = ItemSearchForm(data=QueryDict(query_string="limit=1"))
         self.assertTrue(form.is_valid())
 
-    def test_forms_courses_offset_greater_than_0(self):
+    def test_forms_courses_offset_greater_than_0(self, *_):
         """The `offset` param should be greater than 0."""
         form = ItemSearchForm(data=QueryDict(query_string="offset=-1"))
         self.assertFalse(form.is_valid())
@@ -53,7 +55,7 @@ class ItemSearchFormTestCase(TestCase):
             {"offset": ["Ensure this value is greater than or equal to 0."]},
         )
 
-    def test_forms_courses_offset_integer(self):
+    def test_forms_courses_offset_integer(self, *_):
         """The `offset` param should be an integer."""
         form = ItemSearchForm(data=QueryDict(query_string="offset=a"))
         self.assertFalse(form.is_valid())
@@ -62,7 +64,7 @@ class ItemSearchFormTestCase(TestCase):
         form = ItemSearchForm(data=QueryDict(query_string="offset=1"))
         self.assertTrue(form.is_valid())
 
-    def test_forms_courses_query_between_3_and_100_characters_long(self):
+    def test_forms_courses_query_between_3_and_100_characters_long(self, *_):
         """The `query` param should be between 3 and 100 characters long."""
         form = ItemSearchForm(data=QueryDict(query_string="query=aa"))
         self.assertFalse(form.is_valid())
@@ -88,7 +90,7 @@ class ItemSearchFormTestCase(TestCase):
             {"query": ["Ensure this value has at most 100 characters (it has 101)."]},
         )
 
-    def test_forms_items_single_values_in_querystring(self):
+    def test_forms_items_single_values_in_querystring(self, *_):
         """
         The fields from filter definitions should be normalized as lists. The fields defined
         on the form should be single values (limit, offset and query)
@@ -101,7 +103,7 @@ class ItemSearchFormTestCase(TestCase):
             form.cleaned_data, {"limit": 9, "offset": 3, "query": "maths", "scope": ""}
         )
 
-    def test_forms_items_build_es_query_search_by_match_text(self):
+    def test_forms_items_build_es_query_search_by_match_text(self, *_):
         """
         Happy path: build a query that filters items by matching text
         """
@@ -127,7 +129,7 @@ class ItemSearchFormTestCase(TestCase):
             ),
         )
 
-    def test_forms_items_build_es_query_search_all(self):
+    def test_forms_items_build_es_query_search_all(self, *_):
         """
         Happy path: a match all query is returned
         """